refactor: move cancel
